requirements.txt
# app
streamlit>=1.40,<2
python-dotenv>=1.0,<2
requests>=2.31,<3

# notebooks / data
jupyter>=1.0,<2
ipykernel>=6.29,<7
pandas>=2.0,<3
umap-learn>=0.6.1,<1
streamlit
matplotlib

# hugging face
torch>=2.2,<3
transformers>=4.40,<5
sentence-transformers>=3.0,<4
pillow>=10.0,<12
accelerate>=0.30,<1